Automated Python Script for Data Labeling Cleaning.
I recently wrapped up a project at Prognoz AI, USA as an Associate Data Scientist. Throughout the project, I gained valuable insight into the importance of data cleaning, particularly when dealing with a significant number of categorical variables. As a result, I created an automated solution for label cleaning, here is the code and the explanation of its functionality.